% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{add_prop}
\alias{add_prop}
\title{Add Proportion}
\usage{
add_prop(data, by = "n")
}
\arguments{
\item{data}{A dataframe}

\item{by}{column to use for proportions, defaults to \code{n}.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Adds a proportion column to dataframe
}
\examples{
# ADD_EXAMPLES_HERE
}
