% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{add_source}
\alias{add_source}
\title{Add source column to dataframe}
\usage{
add_source(data, source_name)
}
\arguments{
\item{data}{A dataframe.}

\item{source_name}{Name of the value you want in the new \code{source} column.}
}
\value{
Dataframe
}
\description{
Simply adds a new \code{source} column to the dataframe with the value you specify.
}
\examples{
add_source(mtcars, "R")
}
