% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-archive.R
\name{archive}
\alias{archive}
\title{Archive file or directory}
\usage{
archive(path)
}
\arguments{
\item{path}{Path to file or dir to archive.}
}
\description{
Simple interface to create tarball and save it to archive folder within project.
}
