% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{cmd_check}
\alias{cmd_check}
\title{Checking the command to be passed to Dat system call}
\usage{
cmd_check(x)
}
\arguments{
\item{x}{Command to check}
}
\description{
Checking the command to be passed to Dat system call
}
