% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{create_dat}
\alias{create_dat}
\title{Create Dat metadata}
\usage{
create_dat(dir)
}
\arguments{
\item{dir}{Directory to create Dat.}
}
\description{
Function to create a dat.json interactively, with title and description.
}
\examples{
\dontrun{
  create_dat('.')
  create_dat('data/')
}
}
