% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_create}
\alias{dat_create}
\title{Create a Dat folder in current directory}
\usage{
dat_create()
}
\value{
Dat link
}
\description{
Function to create a dat.json interactively, with title and description.
}
