% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-install.R
\name{dat_install}
\alias{dat_install}
\title{Install R package from Dat link}
\usage{
dat_install(link)
}
\arguments{
\item{link}{Dat link}
}
\value{
Message.
}
\description{
Directly install an R package from a Dat link.
}
\examples{
\dontrun{
  dat_install('dat://datr-chris.hashbase.io')
}
}
