% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_log}
\alias{dat_log}
\title{Show the Dat log}
\usage{
dat_log(link, dir = ".")
}
\arguments{
\item{link}{Dat link (network)}

\item{dir}{Dat folder (local)}
}
\value{
Nothing (for now)
}
\description{
Function to show the log of the Dat folder. You can either retrieve the
log of a Dat link (argument \code{link}) or of a local Dat folder
(argument \code{dir}). Defaults to giving the log of the working directory.
If a Dat link is provided, this overrides the \code{dir} argument.
}
