% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_pull}
\alias{dat_pull}
\title{Update Dat folder continuously}
\usage{
dat_pull(dir = NULL)
}
\arguments{
\item{dir}{Optional, path to Dat directory}
}
\value{
Nothing (for now)
}
\description{
Function to keep updating a Dat folder. This function only receives
updates and does not send changes back. In other words, this is what you
need most often for other people's Dat links.
}
\details{
Note that this function does not exit until you tell R to. Rstudio does
not really like this kind of functionality, but from a terminal you will
see more plus the way to quit.
}
