% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_share}
\alias{dat_share}
\title{Share folder to network}
\usage{
dat_share()
}
\description{
Function to start sharing the current working directory to the Dat
network. Initializes a .dat folder with cryptographic keys. Note that
this is a continuous function and requires the user to abort (does not
work well with Rstudio in general).
}
