% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_sync}
\alias{dat_sync}
\title{Synchronize current folder with Dat network}
\usage{
dat_sync()
}
\description{
Function to resume sharing the folder with the Dat network, if already
created (\code{\link{dat_create}}) locally or cloned
(\code{\link{dat_clone}}).
}
