% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{dat_version}
\alias{dat_version}
\title{Which version of Dat is installed?}
\usage{
dat_version()
}
\value{
Dat version string (if installed)
}
\description{
Which version of Dat is installed?
}
\examples{
dat_version()
}
