% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datr.R
\docType{package}
\name{datr}
\alias{datr}
\alias{datr-package}
\title{datr: Dat Protocol Interface}
\description{
Interface with the 'Dat' p2p network protocol <https://datproject.org>. 
Clone archives from the network, share your own files, and install 
packages from the network.
}
