% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{deregister}
\alias{deregister}
\alias{remove_data}
\alias{remove_tidy}
\alias{remove_raw}
\title{remove data from data registry}
\usage{
remove_data(name, reg_type)

remove_tidy(name)

remove_raw(name)
}
\arguments{
\item{name}{Name of data file.}

\item{reg_type}{Data types to remove: raw, tidy, both.}
}
\description{
This function will delete the named file from the data register, and will
remove the file from disk as well (unless this is turned off)
}
