% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{get_filepath}
\alias{get_filepath}
\title{Get file path of tidy data}
\usage{
get_filepath(name, reg_type)
}
\arguments{
\item{name}{Name of data.}

\item{reg_type}{Whether it is stored in raw or tidy.}
}
\value{
Character vector of absolute filepath.
}
\description{
Function returns the absolute path to a tidy data file.
}
