% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{get_root}
\alias{get_root}
\title{Find data directory}
\usage{
get_root()
}
\value{
root path as character or NULL.
}
\description{
\code{get_root()} recursively searches backwards through the working directories to find
a file named DESCRIPTION. Will try up to 4 times before giving up.
}
