% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{get_root_dir}
\alias{get_root_dir}
\title{Find data directory}
\usage{
get_root_dir()
}
\value{
root path as character or NULL.
}
\description{
\code{get_root_dir()} recursively searches backwards through the working directories to find
a directory named 'data/tidy'. Will try up to 4 times before giving up.
}
