% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_utils.R
\name{input}
\alias{input}
\title{Input helper}
\usage{
input(prompt)
}
\arguments{
\item{prompt}{message to display.}
}
\value{
Text that was inputed
}
\description{
Provides appropriate input function depending on whether session is interactive
}
\examples{
input("What do you want to ask? ")
}
