% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-install.R
\name{install_dat}
\alias{install_dat}
\title{Install Dat from npm}
\usage{
install_dat(os = "windows", pkg = "apt")
}
\arguments{
\item{os}{'windows', 'mac', or 'linux'}

\item{pkg}{Package manager (only for linux; typically 'apt' or 'dnf')}
}
\value{
NULL. Prints stdout of terminal along the way.
}
\description{
Directly install Dat from npm. If node + npm are not yet available, the 
function will first try to install these. NOTE: at the moment the
install_mac() and install_windows() functions have not been tested (TBD).
}
