% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{install_linux}
\alias{install_linux}
\title{Install node on Linux}
\usage{
install_linux(pkg = "apt")
}
\arguments{
\item{pkg}{Package manager of Linux distro (typically \code{apt} or 
\code{dnf})}
}
\value{
Nothing. Prints terminal output.
}
\description{
In order to install Dat, npm needs to be available on the system call. 
This becomes available when nodeJS (\url{https://nodejs.org}) is available.
Using the specified package manager (usually \code{apt} [Debian based] or
\code{dnf} [Redhat based]), this function installs node.
}
\details{
In order to ensure sudo is not needed to install with \code{npm} in 
\code{install_dat()}, the NODE_PATH is exported into the user's .profile
}
