% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{install_node}
\alias{install_node}
\title{Wrap node install functions across OS}
\usage{
install_node(os, pkg)
}
\arguments{
\item{os}{'windows', 'mac', or 'linux'}

\item{pkg}{Package manager [only for 'linux']}
}
\value{
Nothing. Prints terminal output.
}
\description{
Wrap node install functions across OS
}
