% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{install_node_mac}
\alias{install_node_mac}
\title{Install node with brew on Mac}
\usage{
install_node_mac()
}
\value{
Nothing. Prints terminal output.
}
\description{
In order to install Dat, npm needs to be available on the system call. 
This becomes available when nodeJS (\url{https://nodejs.org}) is available.
Using Brew (\url{https://brew.sh/}), this function installs node.
}
\details{
The function wraps a system call that gets an install script and uses Ruby
to install it.
}
