% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{install_windows}
\alias{install_windows}
\title{Install node with Chocolatey on windows}
\usage{
install_windows()
}
\value{
Nothing. Prints terminal output.
}
\description{
In order to install Dat, npm needs to be available on the system call. 
This becomes available when nodeJS (\url{https://nodejs.org}) is available.
Using Chocolatey (\url{https://chocolatey.org/}), this function installs 
node.
}
\details{
The function wraps a system call that gets the install script from
Chocolatey.
}
