% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{load_utils}
\alias{load_utils}
\title{Load utils folder}
\usage{
load_utils()
}
\description{
\code{load_utils()} is another standardisation method from the datr library intended
to help keep data science projects organised and repeatable.
}
\details{
It makes the assumption that there is a code/utils directory, which will then
be sourced.
}
