% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{log_dat}
\alias{log_dat}
\title{Show Dat archive log}
\usage{
log_dat(path)
}
\arguments{
\item{path}{Dat archive, can both be local or on the network.}
}
\value{
Console log.
}
\description{
Retrieve log of the Dat folder. You can either retrieve the
log of a Dat link (argument \code{link}) or of a local Dat folder
(argument \code{dir}). Defaults to giving the log of the working directory.
If a Dat link is provided, this overrides the \code{dir} argument.
}
\examples{
log_dat(path = '.')
log_dat(path = 'dat://pastedat-taravancil.hashbase.io')
}
