% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{model_loaders}
\alias{model_loaders}
\alias{load_outputs}
\alias{load_config}
\alias{load_inputs}
\title{Load model outputs}
\usage{
load_outputs(name, subdir = NULL, ...)

load_config(name, subdir = NULL, ...)

load_inputs(name, subdir = NULL, ...)
}
\arguments{
\item{name}{Name of file to load.}

\item{subdir}{An optional subdir.}

\item{...}{Arguments to pass on to the appropriate file reader.}
}
\description{
\code{load_output()} and \code{load_config()} utilizes the power of a tidy project structure
to load a named file from the model folder.
}
