% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{pull_dat}
\alias{pull_dat}
\title{Update Dat folder}
\usage{
pull_dat(dir = NULL)
}
\arguments{
\item{dir}{Optional, path to Dat directory}
}
\value{
Nothing (for now)
}
\description{
Function to continuously update a Dat folder until abort. This function only
receives updates.
}
