% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{relative_change}
\alias{relative_change}
\title{Calculate relative change}
\usage{
relative_change(before, after)
}
\arguments{
\item{before}{DESCRIPTION.}

\item{after}{DESCRIPTION.}
}
\value{
Numeric vector
}
\description{
\eqn{\frac{after-before}{before}}
}
\examples{
x <- 1:10
y <- x * 2
relative_change(x, y)
}
