% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{rename_data}
\alias{rename_data}
\title{Rename data}
\usage{
rename_data(original_name, new_name, reg_type)
}
\arguments{
\item{original_name}{Current name.}

\item{new_name}{Name to change to.}

\item{reg_type}{Which register to change it in. Either raw, tidy, or both.}
}
\description{
Rename a data file and keep the files organised.
}
