% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{rename_source}
\alias{rename_source}
\title{Rename a source}
\usage{
rename_source(original_name, new_name, reg_type)
}
\arguments{
\item{original_name}{Original source name.}

\item{new_name}{New source name.}

\item{reg_type}{Register type. Options: raw, tidy, both.}
}
\description{
Rename all sources in the register, applied to either raw, tidy, or both.
}
