% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{replace_raw}
\alias{replace_raw}
\title{Replace a data frame in the raw register.}
\usage{
replace_raw(filepath, name)
}
\arguments{
\item{filepath}{Path to replacement file.}

\item{name}{Name of data file to replace.}
}
\description{
Removes the registered data with the same name and replaces it with the specified file.
}
