% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{replace_tidy}
\alias{replace_tidy}
\title{Replace a data frame in the tidy register.}
\usage{
replace_tidy(data, name)
}
\arguments{
\item{data}{A data.frame.}

\item{name}{Name of data file to replace.}
}
\description{
Removes the registered data with the same name and replaces it with the given data.
}
