% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{save_config_text}
\alias{save_config_text}
\title{Save text to model config}
\usage{
save_config_text(text, filename, subdir = NULL)
}
\arguments{
\item{text}{A character vector.}

\item{filename}{Name to use for saved file. Must include a file extension that is
either a delim or feather file.}

\item{subdir}{An optional subdir.}
}
\description{
\code{save_config_text()} utilizes the power of a tidy project structure to save a
character vector to the model config folder.
}
