% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{save_outputs}
\alias{save_outputs}
\title{Save data to model outputs}
\usage{
save_outputs(data, name, subdir = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{name}{Name to use for saved file. \code{.fe} will be added by default.}

\item{subdir}{An optional subdir.}
}
\description{
\code{save_output()} utilizes the power of a tidy project structure to save a data.frame
as a feather file to the model outputs folder.
}
