% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{save_raw}
\alias{save_raw}
\title{Save Raw Data}
\usage{
save_raw(filepath, source, url = NA_character_)
}
\arguments{
\item{filepath}{Path to file to add.}

\item{source}{Source name.}
}
\description{
Registers raw data into project and saves to file structure.
}
