% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-data.R
\name{search_data}
\alias{search_data}
\title{Search the registers}
\usage{
search_data(name)
}
\arguments{
\item{name}{A keyword to search the name fields.}
}
\value{
A tibble.
}
\description{
Look up a key word in both registers
}
