% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-tidy-data.R
\name{search_source}
\alias{search_source}
\title{Search the registers for sources}
\usage{
search_source(source)
}
\arguments{
\item{name}{A keyword to search the source fields.}
}
\value{
A tibble.
}
\description{
Look up a key word in both registers for matching sources
}
