% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{share_dat}
\alias{share_dat}
\title{Share folder to network}
\usage{
share_dat(dir = ".")
}
\arguments{
\item{dir}{Directory to share. Defaults to working directory.}
}
\description{
Share directory to the Dat network, until aborted. Requires write access
to Dat archive.
}
\examples{
share_dat(dir = tempdir())
}
