% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_folder.R
\name{source_folder}
\alias{source_folder}
\title{source_folder}
\usage{
source_folder(folder, silent = FALSE)
}
\arguments{
\item{folder}{the directory to source.}

\item{silent}{logical. If TRUE, information is given about the number of files sourced. Defaults to FALSE}
}
\description{
Recursively source a folder, loading all files with .R ext into the global environment.
}
\examples{
\dontrun{
source_folder("path/to/myfolder")
}
}
