% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{source_helpers}
\alias{source_helpers}
\alias{source_utils}
\alias{source_r}
\alias{source_model}
\alias{source_app}
\alias{source_math}
\alias{source_globals}
\alias{source_functions}
\title{Source Helpers}
\usage{
source_utils(subdir = NULL, ...)

source_r(subdir = NULL, ...)

source_model(subdir = NULL, ...)

source_app(subdir = NULL, ...)

source_math(subdir = NULL, ...)

source_globals()

source_functions(subdir = NULL, ...)
}
\arguments{
\item{subdir}{If required, a subdir can be specified.}

\item{...}{
  Arguments passed on to \code{\link[=source_folder]{source_folder}}
  \describe{
    \item{\code{folder}}{The directory to source.}
    \item{\code{incl}}{An optional regex pattern to use to filter what is included.}
    \item{\code{excl}}{An optional regex pattern to use to exclude files or folders from the search.}
    \item{\code{stop_on_error}}{If \code{TRUE} (default), code will stop on error, with info. If false, only a warning will be issued.}
    \item{\code{silent}}{If \code{TRUE} (default), information is given about the number of files sourced.}
    \item{\code{verbose}}{If true, each file will be printed to console on sourcing.}
  }}
}
\description{
\code{source_xx()} are helper functions that assist with quickly sourcing the content of a
folder with a tidy project structure. Under the hood, these function use \code{source_folder()}.
}
