% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{status_dat}
\alias{status_dat}
\title{Status of the Dat folder}
\usage{
status_dat()
}
\value{
Console log.
}
\description{
Get some information Dat folder in the working directory. Includes: Dat
link, the number of files, and the current version.
}
\examples{
\dontrun{
  status_dat()
}
}
