% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{sync_dat}
\alias{sync_dat}
\title{Synchronize with Dat network}
\usage{
sync_dat()
}
\description{
Synchronize directory with the Dat network. If write access to the Dat is
available, this is equivalent to \code{\link{share_dat}}. If read-only
access is available, this is equivalent to \code{\link{pull_dat}}.
}
\examples{
sync_dat()
}
