% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-archive.R
\name{unarchive}
\alias{unarchive}
\title{Unarchive file or directory}
\usage{
unarchive(name)
}
\arguments{
\item{name}{Name of tarball, sans ".tar.gz".}
}
\description{
Simple helper to extract tarball from project archive and put it back where it was.
}
