% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-archive.R
\name{unarchive_pattern}
\alias{unarchive_pattern}
\title{Unarchive file(s) by regex pattern}
\usage{
unarchive_pattern(pattern)
}
\arguments{
\item{pattern}{Regex pattern to use.}
}
\description{
Simple helper to extract tarball from project archive using regex pattern matching.
}
