% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{update_raw}
\alias{update_raw}
\title{Replace a data frame in the raw register.}
\usage{
update_raw(data = NULL, name, filepath = NULL)
}
\arguments{
\item{data}{A data frame. Either this or a filename.}

\item{name}{Name of data file to replace.}

\item{filepath}{Path to replacement file (Must have either this or a dataframe).}
}
\description{
Removes the registered data with the same name and replaces it with the specified file.
}
