% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{update_tidy}
\alias{update_tidy}
\title{Replace a data frame in the tidy register.}
\usage{
update_tidy(data, name)
}
\arguments{
\item{data}{A data frame.}

\item{name}{Name of data file to replace.}
}
\description{
UPdates the registered data with updated data
}
