% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_dat}
\alias{verify_dat}
\title{Verify Dat link}
\usage{
verify_dat(x)
}
\arguments{
\item{x}{Dat link to verify.}
}
\value{
Boolean TRUE [invisible]
}
\description{
This function verifies a Dat link with a heuristic, checking for a 64 
character hash plus an optional version number. Resolving gateways such
as \url{https://hashbase.io} will be implemented later.
}
