% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_gateway}
\alias{verify_gateway}
\title{Verify Dat gateway}
\usage{
verify_gateway(x)
}
\arguments{
\item{x}{Dat gateway link to verify}
}
\description{
Dat links can either be 64-character hex based, or be used in conjunction
with a gateway (e.g., \url{dat://datr-chris.hashbase.io} for this package)
}
