% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_hash}
\alias{verify_hash}
\title{Verify the hash of a Dat link}
\usage{
verify_hash(x)
}
\arguments{
\item{x}{Dat link}
}
\value{
Boolean TRUE [invisible]
}
\description{
Verify the hash of a Dat link
}
