% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_pipeline}
\alias{verify_pipeline}
\title{Verify Dat link on various aspects}
\usage{
verify_pipeline(x)
}
\arguments{
\item{x}{Dat link to verify.}
}
\value{
Boolean TRUE [invisible]
}
\description{
This function provides a wrapper for the various aspects of a Dat link
such as the protocol (\code{\link{verify_protocol}}), the hash
(\code{\link{verify_hash}}) and the gateway (\code{\link{verify_gateway}};
not yet implemented).
}
