% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_protocol}
\alias{verify_protocol}
\title{Verify the protocol of a Dat link}
\usage{
verify_protocol(x)
}
\arguments{
\item{x}{Dat link}
}
\value{
Boolean TRUE [invisible]
}
\description{
Verify the protocol of a Dat link
}
