% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{verify_version}
\alias{verify_version}
\title{Verify a versioned Dat link}
\usage{
verify_version(x)
}
\arguments{
\item{x}{Dat link}
}
\value{
Boolean TRUE [invisible]
}
\description{
Verify a versioned Dat link
}
