% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{clone_dat}
\alias{clone_dat}
\title{Clone a Dat link to directory}
\usage{
clone_dat(link, dir)
}
\arguments{
\item{link}{Dat link (always clones latest version now)}

\item{dir}{Directory to clone to. If non-existent, will create.}
}
\description{
Function to clone a Dat link to the machine.
}
\examples{
\dontrun{
  clone_dat(link = 'dat://datr-chris.hashbase.io', dir = tempdir())
}
}
