% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{create_dat}
\alias{create_dat}
\title{Create a Dat folder in current directory}
\usage{
create_dat()
}
\value{
Dat link
}
\description{
Function to create a dat.json interactively, with title and description.
}
