% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dat.R
\name{dat_install}
\alias{dat_install}
\title{Install Dat}
\usage{
dat_install(os = "win", ver = "13.10.0")
}
\arguments{
\item{os}{'win' (Windows), 'mac', or 'linux'}

\item{ver}{Version. See \url{https://github.com/datproject/dat/releases}}
}
\value{
NULL. Prints stdout of terminal along the way.
}
\description{
Download the released binaries for Dat and add them to your environment. If
a more recent version has been released, you can specify it as argument (we
try to update it accordingly, but might sometimes have time off!).
}
