% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datr.R
\docType{package}
\name{datr}
\alias{datr}
\alias{datr-package}
\title{datr: Dat Protocol Interface}
\description{
The datr package provides R users with a way to interface with 
the Dat p2p network. Currently, the package allows you to clone Dat archives
from the network, share your own files, and install an R package from a Dat.
}
