% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{deregister}
\alias{deregister}
\title{Deregister data from data registry}
\usage{
deregister(name, rm = TRUE)
}
\arguments{
\item{name}{Name of data file.}

\item{rm}{Logical. Whether to remove the file as well.}
}
\description{
This function will delete the named file from the data register, and will remove the file from disk as well (unless this is turned off)
}
