% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-install.R
\name{install_dat}
\alias{install_dat}
\title{Install R package from Dat link}
\usage{
install_dat(link)
}
\arguments{
\item{link}{Dat link}
}
\value{
Message.
}
\description{
Directly install an R package from a Dat link.
}
\examples{
\dontrun{
  install_dat('dat://datr-chris.hashbase.io')
}
}
