% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{loaders}
\alias{loaders}
\alias{load_clean}
\alias{load_raw}
\title{Load Clean/Raw get your data smartly}
\usage{
load_clean(filename)

load_raw(filename)
}
\arguments{
\item{filename}{Name of filename to extract. File extensions are unneccesary and will be inferred. Use only if filenames are ambiguous.}
}
\value{
A tibble.
}
\description{
These functions assume that your project includes a data/clean and data/raw directory.
They can handle file types of .csv, .tsv, .fe, & excel.
}
\examples{
load_clean("cleanfile")
load_clean("subfolder/to/cleansubfile")
}
