% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise-data.R
\name{loaders}
\alias{loaders}
\alias{read_tidy}
\alias{read_raw}
\alias{read_raw_cells}
\title{Load Tidy/Raw get your data smartly}
\usage{
read_tidy(name, append_source = FALSE, ...)

read_raw(name, append_source = FALSE, ...)

read_raw_cells(name, ...)
}
\arguments{
\item{name}{Name of filename to extract. File extensions are unnecessary and
will be inferred. Use only if filenames are ambiguous.}
}
\value{
A tibble.
}
\description{
These functions assume that your data has been registered into the datrstudio
data management system. If not, do that first. They can handle file types of
.csv, .tsv, .fe, yaml, & excel.
}
\details{
Note that the \code{load_raw_cells()} variant enables an excel type to be loaded
as xlsx_cells for further manipulation (e.g. via the brilliant unpivotr package).
}
