% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{log_dat}
\alias{log_dat}
\title{Show Dat archive log}
\usage{
log_dat(link, dir = ".")
}
\arguments{
\item{link}{Dat link (network)}

\item{dir}{Dat folder (local)}
}
\value{
Console log.
}
\description{
Retrieve log of the Dat folder. You can either retrieve the
log of a Dat link (argument \code{link}) or of a local Dat folder
(argument \code{dir}). Defaults to giving the log of the working directory.
If a Dat link is provided, this overrides the \code{dir} argument.
}
\examples{
\dontrun{
  log_dat()
  log_dat(link = 'dat://pastedat-taravancil.hashbase.io')
}
}
