% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{pull_dat}
\alias{pull_dat}
\title{Update Dat folder}
\usage{
pull_dat()
}
\description{
Function to update working directory if it is a Dat folder. This function
only receives updates.
}
\examples{
pull_dat()
}
