% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-Rmd.R
\name{rmd_setup}
\alias{rmd_setup}
\title{Setup Rmarkdown consistently}
\usage{
rmd_setup(
  subdir = "",
  export = FALSE,
  validation = FALSE,
  figdir = "figures",
  png = TRUE,
  pdf = FALSE,
  echo = FALSE
)
}
\description{
Ensures that Rmarkdown is setup the same, even if the setup itself changes in the future.
}
