% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{save_tidy}
\alias{save_tidy}
\title{Save tidy data in a standardised way}
\usage{
save_tidy(data, filename, force = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{force}{Logical. Whether to prompt before overwrite.}

\item{name}{Filename to save on disk. Can include subdirectories.}
}
\description{
This function standardises the saving of dataframes into the data/tidy dir as a
feather file
}
\details{
This group of functions assumes that you have a \code{ROOT_PATH} and \code{PROJECT_NAME} envir
var set. If these are not found, they'll try to find your data path but may fail.
}
