% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{share_dat}
\alias{share_dat}
\title{Share folder to network}
\usage{
share_dat()
}
\description{
Function to start sharing the current working directory to the Dat
network. Initializes a .dat folder with cryptographic keys. Note that
this is a continuous function and requires the user to abort (does not
work well with Rstudio in general).
}
