% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-folder.R
\name{source_folder}
\alias{source_folder}
\title{Recursively source all R files in a folder and its subfolders}
\usage{
source_folder(
  folder,
  incl = NULL,
  excl = NULL,
  silent = FALSE,
  verbose = FALSE,
  stop_on_error = TRUE
)
}
\arguments{
\item{folder}{The directory to source.}

\item{incl}{An optional regex pattern to use to filter what is included.}

\item{excl}{An optional regex pattern to use to exclude files or folders from the search.}

\item{silent}{If \code{TRUE} (default), information is given about the number of files sourced.}

\item{verbose}{If true, each file will be printed to console on sourcing.}

\item{stop_on_error}{If \code{TRUE} (default), code will stop on error, with info. If false, only a warning will be issued.}
}
\description{
Recursively source a folder, loading all files with .R ext into the global environment.
}
