% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported-tidy-helpers.R
\name{source_helpers}
\alias{source_helpers}
\alias{source_utils}
\alias{source_r}
\alias{source_model}
\title{Source Helpers}
\usage{
source_utils(subdir = "", ...)

source_r(subdir = "", ...)

source_model(subdir = "", ...)
}
\arguments{
\item{subdir}{If required, a subdir can be specified.}

\item{...}{Optional arguments to pass on to source_folder().}
}
\description{
\code{source_utils()}, \code{source_model()}, and \code{source_r()}. Under the hood, these
function use \code{source_folder()}.
}
\details{
These functions assume that a tidy project structure according to datr principles.
}
