% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{status_dat}
\alias{status_dat}
\title{Status of the Dat folder}
\usage{
status_dat(dir = ".")
}
\arguments{
\item{dir}{Location of Dat folder. Defaults to working directory}
}
\value{
Nothing (for now)
}
\description{
Get some information about a Dat folder, including the \code{link}, the
number of files, and the current version.
}
