% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrapper.R
\name{sync_dat}
\alias{sync_dat}
\title{Synchronize current folder with Dat network}
\usage{
sync_dat()
}
\description{
Function to resume sharing the folder with the Dat network, if already
created (\code{\link{create_dat}}) locally or cloned
(\code{\link{clone_dat}}).
}
